moving time inequality to prevent duplicate id on backwards time
If time moves backwards so that ts is before lts, the ts/lts equality check
fails and the sequence number is set to 0. If time moves back to the correct
millisecond on the next request, the sequence number at 0 is used, instead of
whatever sequence number it was previously on.

By moving the going-backwards-time check before the inequality, we avoid this.